# src/app/api/capital-readiness/policy/route.ts
import { NextRequest, NextResponse } from "next/server";
import { requireStaff } from "@/lib/api/auth-helpers";
import { evaluateTransactionAction } from "@/lib/capital-readiness/policy";
import { capitalPolicyRequestSchema } from "@/lib/capital-readiness/validation";

export const dynamic = "force-dynamic";

function json(body: unknown, status = 200) {
  return NextResponse.json(body, {
    status,
    headers: { "Cache-Control": "no-store, max-age=0" },
  });
}

export async function POST(request: NextRequest) {
  const gate = await requireStaff();
  if (!gate.ok) return gate.response;

  let body: unknown;
  try {
    body = await request.json();
  } catch {
    return json({ error: "Invalid JSON" }, 400);
  }

  const parsed = capitalPolicyRequestSchema.safeParse(body);
  if (!parsed.success) {
    return json({ error: "Validation failed", fields: parsed.error.flatten().fieldErrors }, 400);
  }

  const decision = evaluateTransactionAction(parsed.data);
  return json({
    evaluationOnly: true,
    evaluatedBy: gate.session.userId,
    decision,
  });
}